requirements.txt
# AI Agent Identity Governance Platform
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# HashiCorp Vault
hvac==2.3.0

# HTTP Client (async)
aiohttp==3.11.11
httpx==0.28.1

# Machine Learning
scikit-learn==1.6.1
numpy==2.2.1
joblib==1.4.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Dashboard
streamlit==1.41.1
plotly==5.24.1
pandas==2.2.3

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
responses==0.25.6

# Utilities
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0
